Introduced ReadableTokenStore
